+++ HISTORY.md
@@ -306,6 +306,10 @@
 * neural networks library: tweaked "partition" reporter to make sure each class in a data set is adequately represented in both partitions
 * German translation update for newly added "classifier" string
 
+### 2025-10-11
+* objects, threads: new block to prompt for JavaScript, so users,
+  if they choose to allow it, dont have to go into the editor.
+
 ### 2025-10-5
 * blocks and objects: new "object" shape
 * blocks: special case for reportJASMVersion - to put JASMVersion in the